# api/openapi.yaml
openapi: 3.0.3
info:
  title: Symposium API
  version: 1.0.0
  description: |
    Read a Symposium community record, follow it as it changes, and publish through the gate.

    The Symposium API speaks in the specification's model: Artifacts, Objects, properties,
    relationships, Members and addresses (`spec/symposium_specification.md`, profiled by
    `tools/CANONICAL.md`). It is served by the Symposium Data server beside its file-level
    `/v1` API, and reads the community record held there, so every application sees the one
    record under the server's own access rules.

    **The record is read-only here.** Artifacts are immutable. A correction is a new Artifact
    that supersedes the old one, published through `POST /{community}/submissions` and decided
    by the gate. The API offers no update and no delete of a recorded Artifact.

    **Canonical URLs.** Every resource carries a `url`: an absolute URL on the data server that
    names it whichever host asks. An address maps to a URL by a fixed rule (see
    `GET /{community}/resolve` and `api/DESIGN.md`).

    **Freshness.** Every read response carries `position`: the record's cursor and the
    `created` of the newest Artifact the response reflects.

    **Paging.** Every listing whose size grows with the record takes `cursor` and `limit`
    (1 to 1000, default 100) and answers `next`, the cursor of the following page, or `null`
    on the last page. Each listing states its order.

...

...

...

...

...
  license:
    name: MIT
...
servers:
...
security:
  - apiKey: []
paths:
...
  /{community}/record:
    get:
      operationId: getRecord
      tags: [record]
      summary: The record's state
      description: |
        The record's position, its size by type and by Member, its time span, and the
        specification versions its Artifacts declare.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      x-bounded: the specification versions in use, one per revision of the specification
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
      responses:
        '200':
          description: The record's state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts:
    get:
      operationId: listArtifacts
      tags: [record]
      summary: List the record's Artifacts
      description: |
        Every Artifact in the record, as summaries. Ordered by `created` ascending, the order
        the gate accepted them in; `order=-created` reverses it. This listing is also the
        record's change feed: a client that keeps `next` reads each new Artifact once.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
        - name: type
          in: query
          description: Only Artifacts of this `type`. Repeat it to name several.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: published_by
          in: query
          description: Only Artifacts published by this Member's handle.
          schema:
            type: string
            pattern: '^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$'
        - name: created_after
          in: query
          description: Only Artifacts whose `created` is strictly later than this.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: One page of Artifact summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactSummaryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}:
    get:
      operationId: getArtifact
      tags: [record]
      summary: One Artifact, whole
      description: |
        The Artifact in canonical JSON, as the gate stored it, with its canonical URL and the
        links derived from the record: what it supersedes, what supersedes it, and its Member.
        Its address is `@<name>`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
      responses:
        '200':
          description: The Artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/properties/{property}:
    get:
      operationId: getArtifactProperty
      tags: [record]
      summary: One property of an Artifact
      description: |
        The value of one header property. Its address is `@<name>.<property>`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/property'
      responses:
        '200':
          description: The property and its value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/objects:
    get:
      operationId: listObjects
      tags: [record]
      summary: The Objects an Artifact contains
      description: |
        The Artifact's Objects, in the order its canonical JSON lists them. A Model may hold
        many, so the listing is paged.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: type
          in: query
          description: Only Objects of this `type`, such as `Assertion` or `Content`.
          schema:
            type: string
      responses:
        '200':
          description: One page of Objects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/objects/{object}:
    get:
      operationId: getObject
      tags: [record]
      summary: One Object within an Artifact
      description: |
        One Object and the relationships that start or end at it. Its address is
        `@<name>.<object>`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/objectName'
      responses:
        '200':
          description: The Object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/objects/{object}/properties/{property}:
    get:
      operationId: getObjectProperty
      tags: [record]
      summary: One property of an Object
      description: |
        The value of one property of an Object. Its address is `@<name>.<object>.<property>`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/objectName'
        - $ref: '#/components/parameters/property'
      responses:
        '200':
          description: The property and its value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/relationships:
    get:
      operationId: listRelationships
      tags: [record]
      summary: The relationships among an Artifact's Objects
      description: |
        The Artifact's relationships, in the order its canonical JSON lists them. Relationships
        carry no address, so they are reached only through their Artifact.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: rel
          in: query
          description: Only relationships of this kind, such as `grounded_by`.
          schema:
            type: string
      responses:
        '200':
          description: One page of relationships.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/cited-by:
    get:
      operationId: listCitedBy
      tags: [record]
      summary: The later Artifacts that cite this one
      description: |
        Every citation of this Artifact or its content by a later Artifact, ordered by the
        citing Artifact's `created` ascending. `via` narrows it to one kind: the Grounds that
        rest on it, the outputs of an Analysis (`produced_by`), the Analyses that took it as
        input, the Artifacts that supersede it, the Arguments extracted from it, the work that
        serves a ResearchGoal, or citations in prose.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: via
          in: query
          description: Only citations made this way. Repeat it to name several.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CitationVia'
          style: form
          explode: true
      responses:
        '200':
          description: One page of citations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitationPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/supersession:
    get:
      operationId: listSupersession
      tags: [record]
      summary: The Artifact's supersession chain
      description: |
        Every Artifact connected to this one by `supersedes`, in either direction and at any
        distance, ordered by `created` ascending. Each entry says whether it is earlier or later
        and carries its `supersedes_rationale`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: One page of the chain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupersessionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/artifacts/{name}/findings:
    get:
      operationId: listFindings
      tags: [record]
      summary: The validator's findings on an Artifact
      description: |
        The validator's findings, computed by the server. `basis=accepted` gives the findings
        against the record as it stood when the gate accepted the Artifact; they never change.
        `basis=current` gives the findings against the whole record now; a later supersession
        can add one. Ordered by check, then by message.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/artifactName'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: basis
          in: query
          description: Which record the findings are computed against.
          schema:
            type: string
            enum: [accepted, current]
            default: current
        - name: level
          in: query
          description: Only findings at this level.
          schema:
            $ref: '#/components/schemas/FindingLevel'
      responses:
        '200':
          description: One page of findings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/resolve:
    get:
      operationId: resolveAddress
      tags: [record]
      summary: Resolve an address
      description: |
        Resolve any address the specification defines, schema references included, and answer
        what it names, the canonical URL of that thing, and the value where the address reaches
        one. A `text_span`, `csv` or `graph` reference is resolved against the embedded content,
        as the gate verifies it. A `rest` or `download` reference answers its Content Object's
        `location` and `access_method`, unverified.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - name: address
          in: query
          required: true
          description: The address, starting with `@`, URL-encoded.
          schema:
            type: string
            pattern: '^@[^.#]+'
      responses:
        '200':
          description: What the address names.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resolution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /{community}/streams/record:
    get:
      operationId: streamRecord
      tags: [streams]
      summary: Follow the record as the gate accepts
      description: |
        One `artifact` event per Artifact the gate accepts, in `created` order, carrying its
        summary. Each event's `id` is a cursor; reconnecting with `Last-Event-ID` resumes after
        it, so a client misses nothing. A `heartbeat` event every 30 seconds carries the
        position. This is `/symposium sync --watch`.
        The server re-checks the key at every event and every heartbeat, and closes the stream
        when the key is revoked or expired, or its Member leaves the roster.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/lastEventId'
      responses:
        '200':
          description: An open event stream.
          content:
            text/event-stream:
              schema:
                type: array
                description: The stream's events, in order. Each is one SSE frame.
                items:
                  $ref: '#/components/schemas/RecordStreamEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyStreams'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sak_<43 base64url characters>
      description: |
        An API key the admin issued, sent as `Authorization: Bearer sak_…`. The `sak_` prefix
        tells it apart from the data server's Ed25519 tokens (JWTs) and read keys (`sdr_…`),
        which share the same header. The server matches a key by its SHA-256 digest; it
        decrypts nothing to authenticate.
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The server admin's Ed25519 token from `POST /v1/admin/token`, the one the skill already
        holds. The API-key operations accept it alone, so an API key, whatever its scope,
        reads and makes no other key.
  parameters:
    community:
      name: community
      in: path
      required: true
      description: The community's name, matched ignoring case.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_]{1,20}$'
    artifactName:
      name: name
      in: path
      required: true
      description: The Artifact's `name`.
      schema:
        type: string
        pattern: '^[^@.#][^.#]*$'
...
    property:
      name: property
      in: path
      required: true
      description: The property's name.
      schema:
        type: string
        pattern: '^[^@.#][^.#]*$'
...
    cursor:
      name: cursor
      in: query
      description: The `next` of the previous page. Absent for the first page.
      schema:
        type: string
        maxLength: 256
        pattern: '^[A-Za-z0-9_-]+$'
    limit:
      name: limit
      in: query
      description: The most items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    order:
      name: order
      in: query
      description: '`created` for oldest first, `-created` for newest first.'
      schema:
        type: string
        enum: [created, -created]
        default: created
...
  responses:
    BadRequest:
      description: A parameter or the body is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: The credential is absent, unknown, expired or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid and its roles or scope do not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The community or the resource is unknown.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    RecordState:
      type: object
      required: [community, url, position, counts, members, specification_versions]
      properties:
        community:
          $ref: '#/components/schemas/CommunityName'
        url:
          type: string
          format: uri
        position:
          $ref: '#/components/schemas/Position'
        counts:
          type: object
          required: [artifacts, by_type, by_member, findings]
          properties:
            artifacts:
              type: integer
            by_type:
              type: object
              additionalProperties:
                type: integer
            by_member:
              type: object
              additionalProperties:
                type: integer
            findings:
              type: object
              description: Findings as the gate saw each Artifact at its acceptance, by level.
              additionalProperties:
                type: integer
        members:
          type: integer
          description: Handles on the roster.
        first_created:
          type: string
          nullable: true
          format: date-time
        specification_versions:
          type: array
          items:
            type: string
...
    ArtifactSummaryPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ArtifactSummary'
...
    PropertyResource:
      type: object
      required: [url, address, property, value, position]
      properties:
        url:
          type: string
          format: uri
        address:
          $ref: '#/components/schemas/Address'
        property:
          type: string
        value:
          $ref: '#/components/schemas/PropertyValue'
        position:
          $ref: '#/components/schemas/Position'
...
    FindingLevel:
      type: string
      enum: [FAIL, REVIEW]
...