# data-server/service/symposium_api/generated/routers/record.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["record"])


@router.get(
    "/{community}/artifacts",
    response_model=ArtifactSummaryPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listArtifacts",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_artifacts(
    community: str,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    order: Optional[Order] = "created",
    type: Optional[List[str]] = Query(None),
    published_by: Optional[str] = None,
    created_after: Optional[AwareDatetime] = None,
    caller: Caller = Depends(authorize("listArtifacts")),
    service: Service = Depends(service_provider),
) -> ArtifactSummaryPage:
    """List the record's Artifacts"""
    return service.list_artifacts(
        caller,
        community=community,
        cursor=cursor,
        limit=limit,
        order=order,
        type=type,
        published_by=published_by,
        created_after=created_after,
    )


@router.get(
    "/{community}/artifacts/{name}",
    response_model=ArtifactResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getArtifact",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_artifact(
    community: str,
    name: str = ...,
    caller: Caller = Depends(authorize("getArtifact")),
    service: Service = Depends(service_provider),
) -> ArtifactResource:
    """One Artifact, whole"""
    return service.get_artifact(caller, community=community, name=name)


@router.get(
    "/{community}/artifacts/{name}/cited-by",
    response_model=CitationPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listCitedBy",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_cited_by(
    community: str,
    name: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    via: Optional[List[CitationVia]] = Query(None),
    caller: Caller = Depends(authorize("listCitedBy")),
    service: Service = Depends(service_provider),
) -> CitationPage:
    """The later Artifacts that cite this one"""
    return service.list_cited_by(
        caller, community=community, name=name, cursor=cursor, limit=limit, via=via
    )


@router.get(
    "/{community}/artifacts/{name}/findings",
    response_model=FindingPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listFindings",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_findings(
    community: str,
    name: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    basis: Optional[Basis] = "current",
    level: Optional[FindingLevel] = None,
    caller: Caller = Depends(authorize("listFindings")),
    service: Service = Depends(service_provider),
) -> FindingPage:
    """The validator's findings on an Artifact"""
    return service.list_findings(
        caller,
        community=community,
        name=name,
        cursor=cursor,
        limit=limit,
        basis=basis,
        level=level,
    )


@router.get(
    "/{community}/artifacts/{name}/objects",
    response_model=ObjectPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listObjects",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_objects(
    community: str,
    name: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    type: Optional[str] = None,
    caller: Caller = Depends(authorize("listObjects")),
    service: Service = Depends(service_provider),
) -> ObjectPage:
    """The Objects an Artifact contains"""
    return service.list_objects(
        caller, community=community, name=name, cursor=cursor, limit=limit, type=type
    )


@router.get(
    "/{community}/artifacts/{name}/objects/{object}",
    response_model=ObjectResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getObject",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_object(
    community: str,
    name: str = ...,
    object: str = ...,
    caller: Caller = Depends(authorize("getObject")),
    service: Service = Depends(service_provider),
) -> ObjectResource:
    """One Object within an Artifact"""
    return service.get_object(caller, community=community, name=name, object=object)


@router.get(
    "/{community}/artifacts/{name}/objects/{object}/properties/{property}",
    response_model=PropertyResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getObjectProperty",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_object_property(
    community: str,
    name: str = ...,
    object: str = ...,
    property: str = ...,
    caller: Caller = Depends(authorize("getObjectProperty")),
    service: Service = Depends(service_provider),
) -> PropertyResource:
    """One property of an Object"""
    return service.get_object_property(
        caller, community=community, name=name, object=object, property=property
    )


@router.get(
    "/{community}/artifacts/{name}/properties/{property}",
    response_model=PropertyResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getArtifactProperty",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_artifact_property(
    community: str,
    name: str = ...,
    property: str = ...,
    caller: Caller = Depends(authorize("getArtifactProperty")),
    service: Service = Depends(service_provider),
) -> PropertyResource:
    """One property of an Artifact"""
    return service.get_artifact_property(
        caller, community=community, name=name, property=property
    )


@router.get(
    "/{community}/artifacts/{name}/relationships",
    response_model=RelationshipPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listRelationships",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_relationships(
    community: str,
    name: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    rel: Optional[str] = None,
    caller: Caller = Depends(authorize("listRelationships")),
    service: Service = Depends(service_provider),
) -> RelationshipPage:
    """The relationships among an Artifact's Objects"""
    return service.list_relationships(
        caller, community=community, name=name, cursor=cursor, limit=limit, rel=rel
    )


@router.get(
    "/{community}/artifacts/{name}/supersession",
    response_model=SupersessionPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listSupersession",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_supersession(
    community: str,
    name: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    caller: Caller = Depends(authorize("listSupersession")),
    service: Service = Depends(service_provider),
) -> SupersessionPage:
    """The Artifact's supersession chain"""
    return service.list_supersession(
        caller, community=community, name=name, cursor=cursor, limit=limit
    )


@router.get(
    "/{community}/record",
    response_model=RecordState,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getRecord",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_record(
    community: str,
    caller: Caller = Depends(authorize("getRecord")),
    service: Service = Depends(service_provider),
) -> RecordState:
    """The record's state"""
    return service.get_record(caller, community=community)


@router.get(
    "/{community}/resolve",
    response_model=Resolution,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="resolveAddress",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def resolve_address(
    community: str,
    address: str = ...,
    caller: Caller = Depends(authorize("resolveAddress")),
    service: Service = Depends(service_provider),
) -> Resolution:
    """Resolve an address"""
    return service.resolve_address(caller, community=community, address=address)